requirements.txt
autopep8==2.0.1
numpy==1.24.1
Pillow==9.4.0
pycodestyle==2.10.0
tomli==2.0.1
